fixing some weird merge stuff
